RF: Cython version of squash routine & benchmarks
Cythonize squash routine and add benchmarks

Gives benchmark outputs for `fast` (Cython) and `slow` (previous code):

Squashing benchmarks
floats
fast 0.75; slow 1.20
With None: fast 0.76; slow 1.21
With mask: fast 0.76; slow 1.14
Other dtype: fast 0.76; slow 1.14
ints
fast 0.75; slow 1.16
With None: fast 0.74; slow 1.16
With mask: fast 0.76; slow 1.11
Other dtype: fast 0.76; slow 1.11
arrays
fast 1.07; slow 4.30
With None: fast 1.11; slow 4.40
With mask: fast 1.06; slow 4.25
Other dtype: fast 1.05; slow 4.23
scalars
fast 0.73; slow 2.78
With None: fast 0.73; slow 2.82
With mask: fast 0.74; slow 1.27
Other dtype: fast 0.73; slow 1.26
